+++ scripts/cs-active-effects.js
@@ -12,8 +12,10 @@ Hooks.once('init', async function() {
 Hooks.on("renderChatMessage", my_renderChatMessage);
 Hooks.on("renderActiveEffectConfig", ActiveEffectDialog_render);
 
-Hooks.once('init', async function() {
-    Hooks.on("createChatMessage", my_createChatMessage);
+Hooks.once('ready', async function() {
+    if (game.modules.get('dae')?.active) {
+        Hooks.on("createChatMessage", my_createChatMessage);
+    }
 })
 
 // We can't use renderActorSheet and renderItemSheet hooks, because that would prevent the FX tab remaining open
@@ -168,9 +170,10 @@ function my_renderChatMessage(message, html, data) {
     html.find('.chat-card-buttons').prepend(`<a class="transferFX" title="${game.i18n.localize('CSACTIVEEFFECTS.TransferToTarget')}"><i class="fas fa-person-rays"></i></a>`)
     html.find('a.transferFX').on('click', async (ev) => {
         //console.log(`copy effects from ${item.uuid} to`, game.user.targets)
-        if (game.modules.get('dae')?.active)
+        if (game.modules.get('dae')?.active) {
+            console.debug(`CS-AE | Calling DAE to apply effects for '${item.name}'`)
             DAE.doEffects(item, true, game.user.targets)
-        else
+        } else
             await local_transfer_effect(fxlist)
     })
 }
@@ -186,18 +189,26 @@ function my_renderChatMessage(message, html, data) {
  */
 
 function my_createChatMessage(message, options, userId) {
-    if (!game.modules.get('dae')?.active) return;
-
     const itemid = message.flags?.data?.itemID;
-    if (!itemid) return;
+    if (!itemid) {
+        console.debug('CS-AE | No item in chat message')
+        return;
+    }
     // If we can't find the item, then we can't process the button
     const actorid = message.flags.data.actorUuid;
     const item = fromUuidSync(`${actorid}.Item.${itemid}`);
-    if (!item) return;
+    if (!item) {
+        console.debug('CS-AE | Failed to find item on Actor');
+        return;
+    }
     
     // Apply any effects which might transfer on item usage:
     const actortoken = game.canvas.scene.tokens.find(t => actorid.endsWith(t.actorId));
-    if (!actortoken) return;
+    if (!actortoken) {
+        console.debug('CS-AE | Failed to find token for actor')
+        return;
+    }
+    console.debug(`CS-AE | Asking DAE to trigger 'selfEffectAlways' for use of '${item.name}'`)
     DAE.doEffects(item, /*activate*/true, [actortoken], {selfEffects: "selfEffectsAlways"})
 }